+++ eslint-plugin-nimbi-debug/index.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    'no-eager-debug': require('./rules/no-eager-debug.cjs')
+  }
+}