+++ eslint.config.cjs
@@ -16,13 +16,15 @@ module.exports = [
       }
     },
     plugins: {
-      'unused-imports': require('eslint-plugin-unused-imports')
+      'unused-imports': require('eslint-plugin-unused-imports'),
+      'nimbi-debug': require('./eslint-plugin-nimbi-debug/index.cjs')
     },
     rules: {
       'no-unused-vars': ['warn', { argsIgnorePattern: '^_' }],
       'unused-imports/no-unused-imports': 'warn',
       'unused-imports/no-unused-vars': ['warn', { vars: 'all', varsIgnorePattern: '^_', args: 'after-used', argsIgnorePattern: '^_' }],
       'no-console': 'off'
+      , 'nimbi-debug/no-eager-debug': 'warn'
     }
   }
 ]